services/video-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
opencv-python==4.8.1.78
openai-whisper==20231117
ffmpeg-python==0.2.0
numpy==1.24.3
torch==2.1.0
